Conflicts:
	Libs/lib_CMake/android/armeabi-v7a/libcrypto.a
	Libs/lib_CMake/android/armeabi-v7a/libcurl.a
	Libs/lib_CMake/android/armeabi-v7a/libssl.a
	Libs/lib_CMake/android/x86/libcrypto.a
	Libs/lib_CMake/android/x86/libcurl.a
	Libs/lib_CMake/android/x86/libssl.a
	Projects/TestBed/android/jni/Application.mk
	Projects/UnitTests/android/jni/Application.mk
	Sources/CMake/Modules/GlobalVariables.cmake
	Sources/Internal/Autotesting/AutotestingSystemLua.cpp
	Sources/Internal/FileSystem/File.h

Former-commit-id: 59c7ed5d5c29081a42b0a5b28e48f0db189ba52a

+++ Sources/Internal/Logger/Logger.cpp
@@ -347,7 +347,7 @@ bool Logger::CutOldLogFileIfExist(const FilePath& logFile) const
 
     const uint32 sizeToCut = cutLogSize;
 
-    const uint32 fileSize = log->GetSize();
+    const uint32 fileSize = static_cast<uint32>(log->GetSize());
     if (sizeToCut >= fileSize)
     {
         return true; // ok! Have less data than we should to cut.
